requirements.txt
streamlit==1.28.1
pandas==2.1.3
matplotlib==3.8.2
psycopg2-binary>=2.9.0
polars>=0.19.0
numpy==1.24.3 